+++ modules/pensions/app/models/pensions/saved_claim.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Pensions
+  class SavedClaim < ::SavedClaim
+  end
+end
