--- modules/pensions/app/models/saved_claim/saved_claim.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Pensions
-  module SavedClaim
-    class SavedClaim < ::SavedClaim
-    end
-  end
-end
